Only show the Via link on the mobile homepage
When visiting the Hypothesis homepage on mobile devices,
only show the Via link.

The bookmarklet is not practically installable for most
Android/iOS users, although the link is still there below
the laptop if anyone really wants to try.

Chrome extensions are currently not supported on mobile :(
