+++ h/static/styles/home.scss
@@ -41,17 +41,26 @@
     text-align: center;
   }
 
+  $laptop-image-width: 800px;
+  $laptop-image-height: 424px;
+
+  $laptop-screenshot-width: 620px;
+  $laptop-screenshot-height: 365px;
+
+  $laptop-screenshot-left-offset: 90px;
+  $laptop-screenshot-top-offset: 29px;
+
   .home__intro-laptop {
     position: relative;
-    width: 800px;
-    height: 424px;
+    width: $laptop-image-width;
+    height: $laptop-image-height;
     margin-left: auto;
     margin-right: auto;
   }
 
   .home__intro-image {
-    width: 800px;
-    height: 424px;
+    width: 100%;
+    height: 100%;
     display: block;
     margin-left: auto;
     margin-right: auto;
@@ -65,10 +74,10 @@
 
     // screenshot size and positioning chosen to
     // fit in the screen area of the laptop image
-    width: 620px;
-    height: 365px;
-    top: 29px;
-    left: 90px;
+    max-width: $laptop-screenshot-width;
+    height: $laptop-screenshot-height;
+    top: $laptop-screenshot-top-offset;
+    left: $laptop-screenshot-left-offset;
   }
 
   .home__intro {
@@ -134,6 +143,10 @@
     text-align: center;
     margin-left: 10px;
     margin-right: 10px;
+  }
+
+  .installer__alt.is-hidden {
+    display: none;
   }
 
   $proxy-border: 1px solid $color-alto;
@@ -226,8 +239,24 @@
   }
 
   @media only screen and (max-width: $break-tablet) {
+    .home__content {
+      margin-top: 40px;
+    }
+
+    .home__intro-container {
+      margin-top: 180px;
+    }
+
     .home__intro-laptop {
-      display: none;
+      width: $laptop-image-width / 3;
+      height: $laptop-image-height / 3;
+    }
+
+    .home__intro-screenshot {
+      width: $laptop-screenshot-width / 3;
+      height: $laptop-screenshot-height / 3;
+      left: $laptop-screenshot-left-offset / 3;
+      top: $laptop-screenshot-top-offset / 3;
     }
   }
 }
