+++ h/templates/includes/header.html.jinja2
@@ -22,7 +22,10 @@
           <a href="/jobs">We're hiring!</a>
       </div>
       <div class="u-stretch"></div>
-      <a class="nav-bar-toggle js-menu-toggle" data-menu-target=".nav-bar-container" href="#">Menu</a>
+      <a class="nav-bar-toggle js-menu-toggle"
+         data-menu-target=".nav-bar-container"
+         title="Toggle the site navigation menu"
+         href="#">Menu</a>
   </div>
   <div class="nav-bar-container">
       <ul class="nav-bar">
